Add Session integration test and unit tests.
