+++ src/Cassandra.BrokenTests/StatementTests.cs
@@ -39,96 +39,6 @@ namespace Cassandra.Tests
         }
 
         [Test]
-        public void SimpleStatement_Constructor_Positional_Values()
-        {
-            var stmt = new SimpleStatement(Query, 1, "value 2", 3L);
-            CollectionAssert.AreEqual(new object[] { 1, "value 2", 3L }, stmt.QueryValues);
-        }
-
-        [Test]
-        public void SimpleStatement_Bind_Positional_Values()
-        {
-            var stmt = new SimpleStatement(Query).Bind(1, "value 2", 10030L);
-            CollectionAssert.AreEqual(new object[] { 1, "value 2", 10030L }, stmt.QueryValues);
-        }
-
-        [Test]
-        public void SimpleStatement_Constructor_No_Values()
-        {
-            var stmt = new SimpleStatement(Query);
-            Assert.AreEqual(null, stmt.QueryValues);
-        }
-
-        [Test]
-        public void SimpleStatement_Bind_No_Values()
-        {
-            var stmt = new SimpleStatement(Query).Bind();
-            Assert.AreEqual(new object[0], stmt.QueryValues);
-        }
-
-        [Test]
-        public void SimpleStatement_Constructor_Named_Values()
-        {
-            var values = new { Name = "Futurama", Description = "In Stereo where available", Time = DateTimeOffset.Parse("1963-08-28") };
-            var stmt = new SimpleStatement(Query, values);
-            var actualValues = new Dictionary<string, object>();
-            Assert.AreEqual(3, stmt.QueryValueNames.Count);
-            Assert.AreEqual(3, stmt.QueryValues.Length);
-            //Order is not guaranteed
-            for (var i = 0; i < stmt.QueryValueNames.Count; i++)
-            {
-                actualValues[stmt.QueryValueNames[i]] = stmt.QueryValues[i];
-            }
-            //Lowercased
-            Assert.AreEqual(values.Name, actualValues["name"]);
-            Assert.AreEqual(values.Description, actualValues["description"]);
-            Assert.AreEqual(values.Time, actualValues["time"]);
-        }
-
-        [Test]
-        public void SimpleStatement_Constructor_Dictionary_Named_Test()
-        {
-            var valuesDictionary = new Dictionary<string, object>
-            {
-                {"Name", "Futurama"},
-                {"Description", "In Stereo where available"},
-                {"Time", DateTimeOffset.Parse("1963-08-28")}
-            };
-            var stmt = new SimpleStatement(valuesDictionary, Query);
-            var actualValues = new Dictionary<string, object>();
-            Assert.AreEqual(3, stmt.QueryValueNames.Count);
-            Assert.AreEqual(3, stmt.QueryValues.Length);
-            //Order is not guaranteed
-            for (var i = 0; i < stmt.QueryValueNames.Count; i++)
-            {
-                actualValues[stmt.QueryValueNames[i]] = stmt.QueryValues[i];
-            }
-            //Lowercased
-            Assert.AreEqual(valuesDictionary["Name"], actualValues["name"]);
-            Assert.AreEqual(valuesDictionary["Description"], actualValues["description"]);
-            Assert.AreEqual(valuesDictionary["Time"], actualValues["time"]);
-        }
-
-        [Test]
-        public void SimpleStatement_Bind_Named_Values()
-        {
-            var values = new { Name = "Futurama", Description = "In Stereo where available", Time = DateTimeOffset.Parse("1963-08-28") };
-            var stmt = new SimpleStatement(Query).Bind(values);
-            var actualValues = new Dictionary<string, object>();
-            Assert.AreEqual(3, stmt.QueryValueNames.Count);
-            Assert.AreEqual(3, stmt.QueryValues.Length);
-            //Order is not guaranteed
-            for (var i = 0; i < stmt.QueryValueNames.Count; i++)
-            {
-                actualValues[stmt.QueryValueNames[i]] = stmt.QueryValues[i];
-            }
-            //Lowercased
-            Assert.AreEqual(values.Name, actualValues["name"]);
-            Assert.AreEqual(values.Description, actualValues["description"]);
-            Assert.AreEqual(values.Time, actualValues["time"]);
-        }
-
-        [Test]
         public void Statement_SetPagingState_Disables_AutoPage()
         {
             var statement = new SimpleStatement();
