+++ src/Cassandra.Tests/ClusterTests.cs
@@ -1,0 +1,56 @@
+//
+//      Copyright (C) DataStax Inc.
+//
+//   Licensed under the Apache License, Version 2.0 (the "License");
+//   you may not use this file except in compliance with the License.
+//   You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//   Unless required by applicable law or agreed to in writing, software
+//   distributed under the License is distributed on an "AS IS" BASIS,
+//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//   See the License for the specific language governing permissions and
+//   limitations under the License.
+//
+
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Net;
+
+using Cassandra.ExecutionProfiles;
+
+using Moq;
+using NUnit.Framework;
+using Assert = NUnit.Framework.Legacy.ClassicAssert;
+
+namespace Cassandra.Tests
+{
+    [TestFixture]
+    public class ClusterUnitTests
+    {
+
+        [Test]
+        public void ClusterConnectThrowsNoHostAvailable()
+        {
+            var cluster = Cluster.Builder()
+             .AddContactPoint("127.100.100.100")
+             .Build();
+            Assert.Throws<NoHostAvailableException>(() => cluster.Connect());
+            Assert.Throws<NoHostAvailableException>(() => cluster.Connect("sample_ks"));
+        }
+
+        [Test]
+        public void ClusterIsDisposableAfterInitError()
+        {
+            const string ip = "127.100.100.100";
+            var cluster = Cluster.Builder()
+             .AddContactPoint(ip)
+             .Build();
+            Assert.Throws<NoHostAvailableException>(() => cluster.Connect());
+            Assert.DoesNotThrow(cluster.Dispose);
+        }
+    }
+}
